refactor: ChatTableViewCell 네이밍 변경 및 클래스 위치 변경
