+++ Projects/Features/TopicFeature/Sources/Cell/Common/Frame/CommentView.swift
@@ -30,9 +30,7 @@ extension TopicDetailCollectionViewCell {
         }()
         private let contentFrame: UIView = UIView()
         private let blurView: CustomVisualEffectView = CustomVisualEffectView(effect: UIBlurEffect(style: .regular), intensity: 0.06)
-        
-        private let contentCell: ChatTableViewCell = ChatTableViewCell()
-        
+        private let contentCell: RepresentativeCommentView = RepresentativeCommentView()
         private let induceSelectChip: PaddingLabel = {
             let label = PaddingLabel(topBottom: 4, leftRight: 14)
             label.backgroundColor = Color.subPurple
@@ -107,6 +105,46 @@ extension TopicDetailCollectionViewCell {
                 }
             }
         }
+    }
+}
+
+extension TopicDetailCollectionViewCell.CommentView {
+    
+    final class RepresentativeCommentView: BaseView {
+        
+        private let profileImageView: UIImageView = {
+            let imageView = UIImageView()
+            imageView.layer.cornerRadius = 22/2
+            imageView.layer.masksToBounds = true
+            imageView.backgroundColor = UIColor(217)
+            return imageView
+        }()
+        private let contentLabel: UILabel = {
+            let label = UILabel()
+            label.text = "나는 10년 전 과거로 가서 주식..."
+            label.setTypo(Pretendard.regular15)
+            label.textColor = Color.white
+            return label
+        }()
+        
+        override func hierarchy() {
+            addSubviews([profileImageView, contentLabel])
+        }
+        
+        override func layout() {
+            profileImageView.snp.makeConstraints{
+                $0.leading.equalToSuperview().offset(16)
+                $0.centerY.equalToSuperview()
+                $0.width.height.equalTo(22)
+            }
+            contentLabel.snp.makeConstraints{
+                $0.leading.equalTo(profileImageView.snp.trailing).offset(10)
+                $0.top.equalToSuperview().offset(4.5)
+                $0.bottom.equalToSuperview().inset(3.5)
+                $0.trailing.equalToSuperview()
+            }
+        }
+        
     }
 }
